+++ diode_measurement/controller.py
@@ -123,7 +123,13 @@ class Controller(QtCore.QObject):
         self.rpc_params: Cache = Cache()
 
         self.view.setProperty("contentsUrl", "https://github.com/hephy-dd/diode-measurement")
-        self.view.setProperty("about", f"<h3>Diode Measurement</h3><p>IV/CV measurements for silicon sensors.</p><p>Version {__version__}</p><p>Copyright &copy; 2021-2022 <a href=\"https://hephy.at\">HEPHY.at</a></p><p>This software is licensed under the GNU General Public License Version 3.</p>")
+        self.view.setProperty("about", f"""
+            <h3>Diode Measurement</h3>
+            <p>IV/CV measurements for silicon sensors.</p>
+            <p>Version {__version__}</p>
+            <p>This software is licensed under the GNU General Public License Version 3.</p>
+            <p>Copyright &copy; 2021-2022 <a href=\"https://hephy.at\">HEPHY.at</a></p>
+        """)
 
         # Controller
         self.ivPlotsController = IVPlotsController(self)
